plugin-graph: rename pipelines → workflows; update exports and tests; adjust AGENTS docs to prefer local src import within package and public entry for external consumers; fix root AGENTS testing note; build/lint/check/test passing
