+++ TASKS.md
@@ -1,6 +1,6 @@
 # Tasks
 
 - [x] Document usage examples for `crawlGraph` and `exploreGraph` in `@civ7/plugin-graph`.
-- [x] Add CLI integration tests verifying the `crawl` and `explore` commands invoke plugin pipelines.
+- [x] Add CLI integration tests verifying the `crawl` and `explore` commands invoke plugin workflows.
 - [x] Clean up file/directory structure for CLI and plugin-graph packages.
-- [x] Improve error handling and logging in graph plugin pipelines.
+- [x] Improve error handling and logging in graph plugin workflows.
