+++ packages/cli/TESTING.md
@@ -11,7 +11,7 @@ pnpm --filter @civ7/cli test
 ## Current smoke tests
 
 - `expandPath` expands `~` to the user's home directory.
-- `crawl` and `explore` commands delegate to `@civ7/plugin-graph` pipelines.
+- `crawl` and `explore` commands delegate to `@civ7/plugin-graph` workflows.
 
 ## Suggested future tests
