+++ packages/plugins/plugin-graph/TESTING.md
@@ -10,7 +10,7 @@ pnpm -F @civ7/plugin-graph test
 
 ## Current smoke tests
 
-- `crawlGraph` and `exploreGraph` pipelines build graphs and render SVG.
+- `crawlGraph` and `exploreGraph` workflows build graphs and render SVG.
 - Invalid seeds surface descriptive errors and log progress when a logger is provided.
 
 ## Suggested future tests
