+++ packages/plugins/plugin-graph/src/index.ts
@@ -3,4 +3,4 @@ export { buildGraphViewerHtml } from './viewer';
 export { renderSvg } from './render';
 export * as crawler from './crawler';
 export * from './types';
-export * from './pipelines';
+export * from './workflows';
